requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
pyodbc==4.0.39
python-dotenv==1.0.0